Expose schema_dir and query_dir to the CLI

+++ simeon/report/utilities.py
@@ -335,7 +335,10 @@ def extract_table_query(table, query_dir):
     return (''.join(query), ''.join(description))
 
 
-def make_user_info_combo(dirname, outname='user_info_combo.json.gz'):
+def make_user_info_combo(
+    dirname, schema_dir=SCHEMA_DIR,
+    outname='user_info_combo.json.gz'
+):
     """
     Given a course's SQL directory, make a user_info_combo report
 
@@ -353,8 +356,13 @@ def make_user_info_combo(dirname, outname='user_info_combo.json.gz'):
                 '{f} does not exist in the SQL bundle.'.format(f=file_)
             )
     schema_file = os.path.join(
-        SCHEMA_DIR, 'schema_user_info_combo.json'
+        schema_dir, 'schema_user_info_combo.json'
     )
+    if not os.path.exists(schema_file):
+        raise MissingSchemaException(
+            'The schema file {f} does not exist. '
+            'Please provide a valid schema directory.'.format(f=schema_file)
+        )
     with open(schema_file) as sfh:
         schema = json.load(sfh).get('user_info_combo')
     users = dict()
@@ -704,6 +712,7 @@ def make_course_axis(dirname, outname='course_axis.json.gz'):
 
 def make_grades_persistent(
     dirname,
+    schema_dir=SCHEMA_DIR,
     first_outname='grades_persistent.json.gz',
     second_outname='grades_persistent_subsection.json.gz'
 ):
@@ -735,7 +744,13 @@ def make_grades_persistent(
             raise OSError(
                 '{f} does not exist in the SQL bundle.'.format(f=file_)
             )
-        with open(os.path.join(SCHEMA_DIR, schema_file)) as sfh:
+        fschema_file = os.path.join(schema_dir, schema_file)
+        if not os.path.exists(fschema_file):
+            raise MissingSchemaException(
+                'The given schema file {f} does not exist. '
+                'Please provide a valid schema directory.'.format(f=fschema_file)
+            )
+        with open(fschema_file) as sfh:
             sname, _ = os.path.splitext(schema_file)
             sname = sname.replace('schema_', '')
             schema = json.load(sfh).get(sname)
@@ -756,13 +771,19 @@ def make_grades_persistent(
             zh.flush()
 
 
-def make_grading_policy(dirname, outname='grading_policy.json.gz'):
+def make_grading_policy(
+    dirname,
+    schema_dir=SCHEMA_DIR,
+    outname='grading_policy.json.gz'
+):
     """
     Generate a file to be loaded into the grading_policy table
     of the given SQL directory.
 
     :type dirname: str
     :param dirname: Name of a course's directory of SQL files
+    :type schema_dir: str
+    :param schema_dir: Directory where schema files live
     :type outname: str
     :param outname: The filename to give it to the generated report
     :rtype: None
@@ -823,13 +844,19 @@ def _extract_mongo_values(record, key, subkey):
     return None
 
 
-def make_forum_table(dirname, outname='forum.json.gz'):
+def make_forum_table(
+    dirname,
+    schema_dir=SCHEMA_DIR,
+    outname='forum.json.gz'
+):
     """
     Generate a file to load into the forum table
     using the given SQL directory
 
     :type dirname: str
     :param dirname: Name of a course's directory of SQL files
+    :type schema_dir: str
+    :param schema_dir: Directory where schema files live
     :type outname: str
     :param outname: The filename to give it to the generated report
     :rtype: None
@@ -859,8 +886,13 @@ def make_forum_table(dirname, outname='forum.json.gz'):
         ),
     }
     schema_file = os.path.join(
-        SCHEMA_DIR, 'schema_forum.json'
+        schema_dir, 'schema_forum.json'
     )
+    if not os.path.exists(schema_file):
+        raise MissingSchemaException(
+            'The schema file {f} does not exist. '
+            'Please provide a valid schema directory.'.format(f=schema_file)
+        )
     with open(schema_file) as sfh:
         schema = json.load(sfh).get('forum')
     with open(file_) as fh, gzip.open(outname, 'wt') as zh:
@@ -936,13 +968,19 @@ def make_problem_analysis(state, **extras):
     return out
 
 
-def make_student_module(dirname, outname='studentmodule.json.gz'):
+def make_student_module(
+    dirname,
+    schema_dir=SCHEMA_DIR,
+    outname='studentmodule.json.gz'
+):
     """
     Generate files to load into studentmodule and problem_analysis
     using the given SQL directory
 
     :type dirname: str
     :param dirname: Name of a course's directory of SQL files
+    :type schema_dir: str
+    :param schema_dir: Directory where schema files live
     :type outname: str
     :param outname: The filename to give it to the generated report
     :rtype: None
@@ -956,9 +994,9 @@ def make_student_module(dirname, outname='studentmodule.json.gz'):
         raise OSError(
             '{f} does not exist in the SQL bundle.'.format(f=file_)
         )
-    with open(pjoin(SCHEMA_DIR, 'schema_studentmodule.json')) as sfh:
+    with open(pjoin(schema_dir, 'schema_studentmodule.json')) as sfh:
         module_schema = json.load(sfh).get('studentmodule')
-    with open(pjoin(SCHEMA_DIR, 'schema_problem_analysis.json')) as sfh:
+    with open(pjoin(schema_dir, 'schema_problem_analysis.json')) as sfh:
         problem_schema = json.load(sfh).get('problem_analysis')
     prob_cols = ('correct_map', 'student_answers')
     with open(file_, encoding='UTF8', errors='ignore') as fh:
@@ -1017,12 +1055,18 @@ def _default_roles():
     }
 
 
-def make_roles_table(dirname, outname='roles.json.gz'):
+def make_roles_table(
+    dirname,
+    schema_dir=SCHEMA_DIR,
+    outname='roles.json.gz'
+):
     """
     Generate a file to be loaded into the roles table of a dataset
 
     :type dirname: str
     :param dirname: Name of a course's directory of SQL files
+    :type schema_dir: str
+    :param schema_dir: Directory where schema files live
     :type outname: str
     :param outname: The filename to give it to the generated report
     :rtype: None
@@ -1086,6 +1130,7 @@ def make_roles_table(dirname, outname='roles.json.gz'):
 
 def make_sql_tables_seq(
     dirnames, verbose=False, logger=None, fail_fast=False, debug=False,
+    schema_dir=SCHEMA_DIR,
 ):
     """
     Given an iterable of SQL directories, make the SQL tables
@@ -1103,6 +1148,8 @@ def make_sql_tables_seq(
     :param fail_fast: Whether or not to bail after the first error
     :type debug: bool
     :param debug: Show the stacktrace that caused the error
+    :type schema_dir: str
+    :param schema_dir: The directory where schema files live
     :rtype: bool
     :return: True if the files are generated, and False otherwise.
     """
@@ -1119,7 +1166,7 @@ def make_sql_tables_seq(
                 msg = 'Making {f} with files in {d}'
                 logger.info(msg.format(f=tbl, d=dirname))
             try:
-                fn(dirname)
+                fn(dirname=dirname, schema_dir=schema_dir)
             except:
                 _, excp, tb = sys.exc_info()
                 if debug:
@@ -1146,6 +1193,7 @@ def make_sql_tables_seq(
 
 def make_sql_tables_par(
     dirnames, verbose=False, logger=None, fail_fast=False, debug=False,
+    schema_dir=SCHEMA_DIR
 ):
     """
     Given a list of SQL directories, make the SQL tables
@@ -1163,6 +1211,8 @@ def make_sql_tables_par(
     :param fail_fast: Whether or not to bail after the first error
     :type debug: bool
     :param debug: Show the stacktrace that caused the error
+    :type schema_dir: str
+    :param schema_dir: The directory where schema files live
     :rtype: bool
     :return: True if the files are generated, and False otherwise.
     """
@@ -1183,7 +1233,7 @@ def make_sql_tables_par(
                     msg = 'Making {f} with files in {d}'
                     logger.info(msg.format(f=tbl, d=dirname))
                 results[(tbl, dirname)] = pool.apply_async(
-                    fn, args=(dirname,)
+                    fn, args=(dirname, schema_dir)
                 )
         fails = []
         for (tbl, dirname), result in results.items():
@@ -1216,8 +1266,8 @@ def make_sql_tables_par(
 
 def make_table_from_sql(
     table, course_id, client, project, append=False,
-    query_dir=QUERY_DIR, wait=False,
-    geo_table='geocode.geoip', youtube_table='videos.youtube',
+    query_dir=QUERY_DIR, schema_dir=SCHEMA_DIR,
+    wait=False, geo_table='geocode.geoip', youtube_table='videos.youtube',
 ):
     """
     Generate a BigQuery table using the given table name,
@@ -1251,9 +1301,10 @@ def make_table_from_sql(
         course_id, 'log', project
     )
     try:
-        cols=',\n'.join(
-            uputils.sqlify_bq_field(f) for f in uputils.get_bq_schema(table)
-        )
+        cols = []
+        for f in uputils.get_bq_schema(table, schema_dir):
+            cols.append(uputils.sqlify_bq_field(f))
+        cols = ',\n'.join(cols)
     except MissingSchemaException:
         cols = ''
     query, description = extract_table_query(table, query_dir)
